take in service name in all commands and removed local breadcrumbs
the service name is now taken in as an optional parameter in existing commands (defaulting to the existing associated service). Also, the local breadcrumbs inside of the .git/ directory where association happened are no longer dropped. These caused confusion when multiple environments had been associated. Now you have to specify an environment with -E, depend on the default, or are prompted to choose